Pack and read GameCube disc images.
This was a huge amount of work. I don't know if I'm using it wrong or what, but there seem to be alignment bugs in libiso9660 and what might have been heap corruption. The libogc DVD code was even weirder, completing reads at the wrong times (?), acknowledging a read while the previous read's data was in the buffer. I wrote my own. And then moved from ISO9660 to GCM.